server/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==24.2
Pillow==10.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
faker==24.0.0
redis==5.0.1
django-redis==5.4.0
